Метод по расписанию перенес в отдельный класс
Оставшиеся вопросы:
Почему в текст принимает только русский язык?
